resources/requirements.txt
requests
google-ads
google-api-python-client
google-auth-oauthlib
slugify
python-slugify
openai